refactor(control): split reload connectivity
